# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for matchmaking, game management, user management, and ELO rating.",
        "title": "Elo Matchmaking Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/lobby/find": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns lobbies for the given game, optionally filtered to those whose Tags contain ALL the comma-separated tags in the query.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lobby"
                ],
                "summary": "List open lobbies for a game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game UUID",
                        "name": "gameID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated tags; lobby must include every tag to be returned",
                        "name": "tags",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lobbies",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/match/game/{gameID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of matches for a specific game",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Get matches for a game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game UUID",
                        "name": "gameID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "matches, nextPage",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/match/join": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upgrades to a WebSocket connection and joins the matchmaking queue for a game. Sends status updates until a match is found.",
                "tags": [
                    "Matchmaking"
                ],
                "summary": "Join matchmaking queue (WebSocket)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game UUID to queue for",
                        "name": "gameID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional sub-queue key (only honored when game.metadata_enabled=true; capped at 4 KB)",
                        "name": "metadata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "JWT token (alternative to Authorization header)",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/match/size": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the number of players currently in the matchmaking queue for a game",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matchmaking"
                ],
                "summary": "Get matchmaking queue size",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game UUID",
                        "name": "gameID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sub-queue key (only honored when game.metadata_enabled=true)",
                        "name": "metadata",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "players_in_queue",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/match/{matchID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns match details. User must be a participant or game owner.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Get a match by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match UUID",
                        "name": "matchID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.MatchResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/matches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of all matches. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "List all matches (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "matches, nextPage",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "internal": {
                    "description": "Stores the error returned by an external dependency"
                },
                "message": {}
            }
        },
...
        "github_com_andy98725_elo-service_src_models.MatchResp": {
            "type": "object",
            "properties": {
                "game_id": {
                    "type": "string"
                },
                "guest_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.UserResp"
                    }
                },
                "server_address": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_andy98725_elo-service_src_models.UserResp": {
            "type": "object",
            "properties": {
                "can_create_game": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Prefix with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}